update blog editor

+++ src/main/java/com/zbnetwork/blog/app/service/UserService.java
@@ -32,5 +32,5 @@ public interface UserService {
     /**
      * find user exists
      */
-    int findUserExists(String username);
+    long findUserExists(String username);
 }
